chore: move v210 output conversion to common files
