+++ src/common/CMakeLists.txt
@@ -12,6 +12,7 @@ set(SOURCES
 		log.cpp
 		tweener.cpp
 		utf.cpp
+		v210.cpp
 )
 if (MSVC)
 	list(APPEND SOURCES
@@ -58,6 +59,7 @@ set(HEADERS
 		timer.h
 		tweener.h
 		utf.h
+		v210.h
 )
 
 casparcg_add_library(common SOURCES ${SOURCES} ${HEADERS})
